+++ src/main/java/ucan/edu/sistema_academico/entities/Disciplina.java
@@ -4,8 +4,6 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.util.List;
-
 @Setter
 @Getter
 @AllArgsConstructor
@@ -18,7 +16,7 @@ public class Disciplina {
     @Id
     @Column(name = "pk_disciplina")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long pkDisciplina;
+    private Integer pkDisciplina;
 
     @Column(name = "nome", nullable = false)
     private String nome;
